requirements.txt
flask
line-bot-sdk
google-generativeai